% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_no_na}
\alias{.check_no_na}
\title{Check for missing values in training data}
\usage{
.check_no_na(data, cols)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{cols}{A \code{character} vector of column names to check in \code{data}.}
}
\value{
Invisibly returns \code{TRUE} if no \code{NA} is found. Otherwise an error is thrown.
}
\description{
Ensures there are no \code{NA} values in any of the relevant columns of the dataset.
}
\keyword{internal}
