% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loss_from_objective}
\alias{loss_from_objective}
\title{Backward and fast path micro evaluator adaptor}
\usage{
loss_from_objective(global_objective_fn)
}
\arguments{
\item{global_objective_fn}{A \code{function} with signature \code{function(D) -> numeric}
that returns a scalar to be minimized. For example \code{\link{objective_default}}.}
}
\value{
A \code{function} \code{loss_fn(val, indices, D)} suitable for use in
\code{\link{ROOT}} and \code{\link{split_node}}. It sets \code{w = val} on \code{indices}
without mutation of the original \code{D} and then returns \code{global_objective_fn(D)}.
}
\description{
Wraps a global objective \code{global_objective_fn(D)} into a splitter compatible
loss function \code{loss_fn(val, indices, D)} by calling \code{\link{objective_if}}
on a temporary copy of \code{D}.
}
