% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{objective_default}
\alias{objective_default}
\title{Generic objective interface}
\usage{
objective_default(D)
}
\arguments{
\item{D}{A \code{data.frame} with at least numeric columns \code{vsq} and \code{w}.}
}
\value{
A \code{numeric(1)} objective value. Returns \code{Inf} when undefined.
}
\description{
Default objective that serves as a proxy for standard error in weighted transported
average treatment effect (WTATE) and population average treatment effect (PATE).
}
\details{
Computes \code{sqrt(sum(vsq_i * w_i)) / (sum(w_i))^2}.
Requires columns \code{vsq} and \code{w} in \code{D}. The goal is to minimize the value.
Supply your own \code{function(D) -> numeric} to use a different objective.
}
