% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{train}
\alias{train}
\title{Train nuisance models for weighting}
\usage{
train(training_data, outcome, treatment, sample)
}
\arguments{
\item{training_data}{A \code{data.frame} that contains the training dataset.}

\item{outcome}{A \code{character(1)} name of the outcome column. Typically this is the observed outcome such as \code{"Yobs"}.}

\item{treatment}{A \code{character(1)} name of the treatment indicator column such as \code{"Tr"}.}

\item{sample}{A \code{character(1)} name of the sample inclusion indicator column such as \code{"S"}.}
}
\value{
A \code{list} with:
\item{pi}{\code{numeric(1)} prevalence of \code{S == 1} in the training data.}
\item{pi_m}{\code{glm} model with binomial family for \eqn{P(S=1 \mid X)}.}
\item{e_m}{\code{glm} model with binomial family for \eqn{P(Tr=1 \mid X, S=1)}.}
}
\description{
Fits models to estimate sampling and treatment propensities on training data by logistic regression.
}
