% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{midpoint}
\alias{midpoint}
\title{Compute the midpoint of a numeric vector}
\usage{
midpoint(X)
}
\arguments{
\item{X}{A numeric vector.}
}
\value{
A numeric scalar giving the midpoint of the finite values in \code{X}. If \code{X} is empty or has no finite values, \code{NA} is returned.
}
\description{
Calculates the midpoint defined as \eqn{( \max(X) + \min(X) ) / 2}, ignoring any NA values.
}
