% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{objective_if}
\alias{objective_if}
\title{Helper: evaluate objective after a hypothetical local change}
\usage{
objective_if(val, indices, D, global_objective_fn)
}
\arguments{
\item{val}{0/1 assignment to apply}

\item{indices}{integer or rownames to receive \code{val}}

\item{D}{data.frame used by \code{global_objective_fn}}

\item{global_objective_fn}{function(D)->scalar}
}
\value{
numeric scalar objective after the hypothetical change
}
\description{
Helper: evaluate objective after a hypothetical local change
}
