% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsROOT.R
\name{print.ROOT}
\alias{print.ROOT}
\title{Print a ROOT fit}
\usage{
\method{print}{ROOT}(x, ...)
}
\arguments{
\item{x}{A \code{"ROOT"} S3 object returned by \code{ROOT()}.}

\item{...}{Currently unused and included for S3 compatibility.}
}
\value{
\code{object} returned invisibly. Printed output is for inspection.
}
\description{
Provides a human-readable brief summary of a \code{ROOT} object, including:
\enumerate{
\item the summary characterization tree \code{f},
\item in generalizability mode (\code{generalizability_path = TRUE}), the
unweighted and weighted estimands with their standard errors
and an explanatory note for the weighted standard error (SE).
}
}
\details{
When \code{generalizability_path = TRUE}, the unweighted estimand corresponds
to a SATE-type quantity and the weighted estimand to a WTATE-type
quantity for the transported target population. When \code{generalizability_path = FALSE},
ROOT is used for general functional optimization and no causal labels
are imposed.
}
\examples{
\dontrun{
ROOT.output = ROOT(diabetes_data,generalizability_path = TRUE, seed = 123)
print(ROOT.output)
}
}
